pyproject.toml
[tool.poetry]
name = "django-project-time-tracker"
version = "0.1.0"
description = "A DJango Rest Framework basic app to handle different projects and time log"
authors = ["dfm88 <diegof.university@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "django_project_time_tracker"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
django-rest-framework = "^0.1.0"
ipdb = "^0.13.9"
psycopg2 = "^2.9.4"
python-dotenv = "^0.21.0"
django-redis = "^5.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"
django-silk = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
